requirements.txt
pytest==9.0.1
requests==2.32.5
PyYAML==6.0.3
python-dotenv==1.2.1
Faker==38.0.0
allure-pytest==2.15.0
